third_party/blink/web_tests/virtual/disable-ua-client-hint/external/wpt/client-hints/accept-ch.wildcard.https.sub-expected.txt

This is a testharness.js-based test.
FAIL Accept-CH with wildcard policy and iframe subresource assert_equals: expected "HadPlatformVersion" but got "MissingPlatformVersion"
Harness: the test ran to completion.
